ZGO: Map FGR to file number instead of file name
The FILES() array is subscripted by the $QS() components of a global
node.  Data appear in nodes of FILES corresponding to the File Global
Root of a FileMan file.  Store the file number there instead of the file
name.  This will allow additional information about the file to be
looked up later.

Change-Id: Ia17b5f08a3e16b5d3acf73f2b1c564f832784e2f