ZGO: Backport changes from master to FOIA import branch
This routine is used during the FOIA VistA extraction and import
process.  Make it easier to use the latest version of the routine by
maintaining it on the FOIA import branch.  Backport changes from
commits:

 22923041 VAF53:  Add DISV to global exclude list
 9f0621fc ZGO: Remove trailing 'DUZ(2)' subscript from roots
 821b3c53 ZGO: Warn and skip non-canonical file global roots
 696781c5 ZGO: Port to GT.M
 8b427a84 EHN:Ignore ^XTMP & ^%ZOSF when extracting globals

and base them on the last FOIA import.  New development on the routine
should start on the import branch from now on.

Change-Id: If77af78040fe2335eacd45dc7bd739ff0796b7a5